backend/transcriber_api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
websockets==12.0
openai-whisper==20231117
python-jose==3.3.0
pydantic==2.5.1
pydantic-settings==2.1.0
fpdf==1.7.2
numpy>=1.22,<1.27
numba==0.58.1
